add mot colors
